+++ workspaces/marketplace/packages/app/src/App.tsx
@@ -52,7 +52,10 @@ import { githubAuthApiRef } from '@backstage/core-plugin-api';
 
 import { getAllThemes } from '@red-hat-developer-hub/backstage-plugin-theme';
 
-import { DynamicMarketplacePluginRouter as Marketplace } from '@red-hat-developer-hub/backstage-plugin-marketplace';
+import {
+  InstallationContextProvider,
+  DynamicMarketplacePluginRouter as Marketplace,
+} from '@red-hat-developer-hub/backstage-plugin-marketplace';
 
 import { apis } from './apis';
 import { entityPage } from './components/catalog/EntityPage';
@@ -133,7 +136,14 @@ const routes = (
     </Route>
     <Route path="/settings" element={<UserSettingsPage />} />
     <Route path="/catalog-graph" element={<CatalogGraphPage />} />
-    <Route path="/extensions" element={<Marketplace />} />
+    <Route
+      path="/extensions"
+      element={
+        <InstallationContextProvider>
+          <Marketplace />
+        </InstallationContextProvider>
+      }
+    />
   </FlatRoutes>
 );