+++ workspaces/marketplace/plugins/marketplace/src/components/MarketplacePluginContent.test.tsx
@@ -30,6 +30,8 @@ import { useExtensionsConfiguration } from '../hooks/useExtensionsConfiguration'
 import { useNodeEnvironment } from '../hooks/useNodeEnvironment';
 import { errorApiRef } from '@backstage/core-plugin-api';
 import { translationApiRef } from '@backstage/core-plugin-api/alpha';
+import { QueryClientProvider } from '@tanstack/react-query';
+import { queryClient } from '../queryclient';
 
 jest.mock('@backstage/core-plugin-api', () => {
   const actual = jest.requireActual('@backstage/core-plugin-api');
@@ -69,6 +71,10 @@ const usePluginPackagesMock = usePluginPackages as jest.Mock;
 const usePluginConfigurationPermissionsMock =
   usePluginConfigurationPermissions as jest.Mock;
 const useExtensionsConfigurationMock = useExtensionsConfiguration as jest.Mock;
+
+queryClient.setDefaultOptions({
+  queries: { retry: false },
+});
 
 beforeEach(() => {
   usePluginConfigurationPermissionsMock.mockReturnValue({
@@ -110,7 +116,9 @@ const renderWithProviders = (ui: ReactNode) =>
         [translationApiRef, mockApis.translation()],
       ]}
     >
-      <BrowserRouter>{ui}</BrowserRouter>
+      <BrowserRouter>
+        <QueryClientProvider client={queryClient}>{ui}</QueryClientProvider>
+      </BrowserRouter>
     </TestApiProvider>,
   );
 
@@ -160,7 +168,7 @@ describe('MarketplacePluginContent', () => {
   });
   it('should have the Install button enabled', async () => {
     const { getByText } = renderWithProviders(
-      <MarketplacePluginContent plugin={plugin} enableActionsButtonFeature />,
+      <MarketplacePluginContent plugin={plugin} />,
     );
     expect(getByText('Install')).toBeInTheDocument();
     const installButton = getByText('Install');
@@ -179,7 +187,9 @@ describe('MarketplacePluginContent', () => {
     });
 
     const { getByText } = renderWithProviders(
-      <MarketplacePluginContent plugin={plugin} enableActionsButtonFeature />,
+      <QueryClientProvider client={queryClient}>
+        <MarketplacePluginContent plugin={plugin} />,
+      </QueryClientProvider>,
     );
     expect(getByText('View')).toBeInTheDocument();
   });
@@ -192,7 +202,7 @@ describe('MarketplacePluginContent', () => {
     });
 
     const { getByText } = renderWithProviders(
-      <MarketplacePluginContent plugin={plugin} enableActionsButtonFeature />,
+      <MarketplacePluginContent plugin={plugin} />,
     );
     expect(getByText('View')).toBeInTheDocument();
   });
@@ -209,7 +219,7 @@ describe('MarketplacePluginContent', () => {
     });
 
     const { getByText } = renderWithProviders(
-      <MarketplacePluginContent plugin={plugin} enableActionsButtonFeature />,
+      <MarketplacePluginContent plugin={plugin} />,
     );
     expect(getByText('Install')).toBeInTheDocument();
     const installButton = getByText('Install');
@@ -226,16 +236,39 @@ describe('MarketplacePluginContent', () => {
     };
 
     const { getByText, getByTestId } = renderWithProviders(
-      <MarketplacePluginContent
-        plugin={installedPlugin}
-        enableActionsButtonFeature
-      />,
+      <MarketplacePluginContent plugin={installedPlugin} />,
     );
     expect(getByText('Actions')).toBeInTheDocument();
     const actionsButton = getByTestId('plugin-actions');
     fireEvent.click(actionsButton);
 
     expect(getByTestId('actions-button')).toBeInTheDocument();
-    expect(getByTestId('enable-plugin')).toBeInTheDocument();
+    expect(getByTestId('disable-plugin')).toBeInTheDocument();
+  });
+
+  it('should have the View button when user has read only access even when the plugin is installed', async () => {
+    usePluginConfigurationPermissionsMock.mockReturnValue({
+      data: {
+        write: 'DENY',
+        read: 'ALLOW',
+      },
+      isLoading: false,
+      error: null,
+      refetch: jest.fn(),
+    });
+    const installedPlugin = {
+      ...plugin,
+      spec: {
+        ...plugin.spec,
+        installStatus: MarketplacePluginInstallStatus.Installed,
+      },
+    };
+
+    const { getByText } = renderWithProviders(
+      <QueryClientProvider client={queryClient}>
+        <MarketplacePluginContent plugin={installedPlugin} />,
+      </QueryClientProvider>,
+    );
+    expect(getByText('View')).toBeInTheDocument();
   });
 });