+++ workspaces/lightspeed/packages/backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # backend
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [0e00cb6]
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-backend@0.5.6
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [2ba3742]
+- Updated dependencies [a79f849]
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-backend@0.5.5
+
 ## 0.0.14
 
 ### Patch Changes