+++ workspaces/marketplace/plugins/marketplace/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @red-hat-developer-hub/backstage-plugin-marketplace
 
+## 0.8.1
+
+### Patch Changes
+
+- 545c7e7: Updated dependency `@mui/styles` to `5.18.0`.
+- c399ade: hide actions menu in view only mode
+
+## 0.8.0
+
+### Minor Changes
+
+- 7aac60c: Introduces `PluginInstallStatusProcessor` to add `spec.installStatus` for Marketplace plugins. Introduces new `installStatus` values: `MarketplacePackageInstallStatus.Disabled` and `MarketplacePluginInstallStatus.Disabled`.
+
+### Patch Changes
+
+- 5ee083d: Fix plugin action button logic to correctly show the "Actions" dropdown when install status is `Installed` or `UpdateAvailable`.
+- dc6d0fb: notify users when plugin installation is successful
+- Updated dependencies [f70ccb1]
+- Updated dependencies [7aac60c]
+  - @red-hat-developer-hub/backstage-plugin-marketplace-common@0.7.0
+
 ## 0.7.0
 
 ### Minor Changes